requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
jinja2==3.1.6
httpx==0.28.1
python-dotenv==1.1.0
urllib3<2
volcengine-python-sdk==5.0.23
